+++ core/src/main/java/feign/Response.java
@@ -18,12 +18,16 @@ package feign;
 import static feign.Util.UTF_8;
 import static feign.Util.checkNotNull;
 import static feign.Util.checkState;
+import static feign.Util.decodeOrDefault;
 import static feign.Util.valuesOrEmpty;
 
+import java.io.ByteArrayInputStream;
 import java.io.Closeable;
 import java.io.IOException;
+import java.io.InputStream;
+import java.io.InputStreamReader;
 import java.io.Reader;
-import java.io.StringReader;
+import java.nio.charset.Charset;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.LinkedHashMap;
@@ -40,14 +44,23 @@ public final class Response {
       int status,
       String reason,
       Map<String, Collection<String>> headers,
-      Reader chars,
+      InputStream inputStream,
       Integer length) {
-    return new Response(status, reason, headers, ReaderBody.orNull(chars, length));
+    return new Response(status, reason, headers, InputStreamBody.orNull(inputStream, length));
   }
 
   public static Response create(
-      int status, String reason, Map<String, Collection<String>> headers, String chars) {
-    return new Response(status, reason, headers, StringBody.orNull(chars));
+      int status, String reason, Map<String, Collection<String>> headers, byte[] data) {
+    return new Response(status, reason, headers, ByteArrayBody.orNull(data));
+  }
+
+  public static Response create(
+      int status,
+      String reason,
+      Map<String, Collection<String>> headers,
+      String text,
+      Charset charset) {
+    return new Response(status, reason, headers, ByteArrayBody.orNull(text, charset));
   }
 
   private Response(int status, String reason, Map<String, Collection<String>> headers, Body body) {
@@ -95,24 +108,29 @@ public final class Response {
      */
     Integer length();
 
-    /** True if {@link #asReader()} can be called more than once. */
+    /** True if {@link #asInputStream()} and {@link #asReader()} can be called more than once. */
     boolean isRepeatable();
+
+    /** It is the responsibility of the caller to close the stream. */
+    InputStream asInputStream() throws IOException;
 
     /** It is the responsibility of the caller to close the stream. */
     Reader asReader() throws IOException;
   }
 
-  private static final class ReaderBody implements Response.Body {
-    private static Body orNull(Reader chars, Integer length) {
-      if (chars == null) return null;
-      return new ReaderBody(chars, length);
+  private static final class InputStreamBody implements Response.Body {
+    private static Body orNull(InputStream inputStream, Integer length) {
+      if (inputStream == null) {
+        return null;
+      }
+      return new InputStreamBody(inputStream, length);
     }
 
-    private final Reader chars;
+    private final InputStream inputStream;
     private final Integer length;
 
-    private ReaderBody(Reader chars, Integer length) {
-      this.chars = chars;
+    private InputStreamBody(InputStream inputStream, Integer length) {
+      this.inputStream = inputStream;
       this.length = length;
     }
 
@@ -127,36 +145,46 @@ public final class Response {
     }
 
     @Override
+    public InputStream asInputStream() throws IOException {
+      return inputStream;
+    }
+
+    @Override
     public Reader asReader() throws IOException {
-      return chars;
+      return new InputStreamReader(inputStream, UTF_8);
     }
 
     @Override
     public void close() throws IOException {
-      chars.close();
+      inputStream.close();
     }
   }
 
-  private static final class StringBody implements Response.Body {
-    private static Body orNull(String chars) {
-      if (chars == null) return null;
-      return new StringBody(chars);
+  private static final class ByteArrayBody implements Response.Body {
+    private static Body orNull(byte[] data) {
+      if (data == null) {
+        return null;
+      }
+      return new ByteArrayBody(data);
     }
 
-    private final String chars;
-
-    public StringBody(String chars) {
-      this.chars = chars;
+    private static Body orNull(String text, Charset charset) {
+      if (text == null) {
+        return null;
+      }
+      checkNotNull(charset, "charset");
+      return new ByteArrayBody(text.getBytes(charset));
     }
 
-    private volatile Integer length;
+    private final byte[] data;
+
+    public ByteArrayBody(byte[] data) {
+      this.data = data;
+    }
 
     @Override
     public Integer length() {
-      if (length == null) {
-        length = chars.getBytes(UTF_8).length;
-      }
-      return length;
+      return data.length;
     }
 
     @Override
@@ -165,16 +193,22 @@ public final class Response {
     }
 
     @Override
-    public Reader asReader() throws IOException {
-      return new StringReader(chars);
-    }
-
-    public String toString() {
-      return chars;
+    public InputStream asInputStream() throws IOException {
+      return new ByteArrayInputStream(data);
     }
 
     @Override
-    public void close() {}
+    public Reader asReader() throws IOException {
+      return new InputStreamReader(asInputStream(), UTF_8);
+    }
+
+    @Override
+    public void close() throws IOException {}
+
+    @Override
+    public String toString() {
+      return decodeOrDefault(data, UTF_8, "Binary data");
+    }
   }
 
   @Override